updated doc

+++ A8/src/binarySearchTree/BinarySearchTreeI.java
@@ -4,5 +4,5 @@ public interface BinarySearchTreeI {
     NodeI addData(Integer data);
     void printTree(PrintVariant pv);
     void updateSum();
-    int getSum(int m, int M);
+    long getSum(int m, int M);
 }
